backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langgraph==0.1.1
langchain==0.2.5
langchain-openai==0.1.8
langchain-community==0.2.5
langchain-groq==0.1.8
pydantic==2.7.4
pydantic-settings==2.10.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
structlog==24.2.0
python-dotenv==1.0.1
tenacity==8.4.1  # For retries
